refactor(sprint-3): update README, refactoring
