+++ src/core/api/HTTPTransport/http.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import HTTPTransport from './HTTPTransport';
+import HTTPTransport from './index';
 
 const xhr = new HTTPTransport();
