+++ src/core/api/index.ts
@@ -1,0 +1,8 @@
+export { default } from './HTTPTransport';
+export { default as chatsAPI } from './chat-api';
+export { default as userAPI } from './user-api';
+export { default as authAPI } from './auth-api';
+export { default as ChatInstance } from './ChatInstance';
+
+export type { ChatData } from './ChatInstance';
+export type { MessageProps } from './ChatInstance';
