--- src/core/http/index.ts
@@ -1,8 +1,0 @@
-export { default } from './HTTPTransport';
-export { default as chatsAPI } from './api/chat-api';
-export { default as userAPI } from './api/user-api';
-export { default as authAPI } from './api/auth-api';
-export { default as ChatInstance } from './api/ChatInstance';
-
-export type { ChatData } from './api/ChatInstance';
-export type { MessageProps } from './api/ChatInstance';
